# docs/code/openapi.yaml
openapi: 3.1.0
info:
  title: neyialiyorlar Financial Metrics API
  version: v1
  description: |
    REST + WebSocket API for real-time financial metrics and dashboard data.

    **Contract rules:**
    - Every response includes a `flag` field (required): "fresh" | "stale" | "approx"
    - `value` field is nullable (number | null); gap rows use null, never omitted
    - All timestamps are ISO 8601 UTC (e.g. 2026-06-14T15:00:00Z)
    - Versioned under /api/v1/; future breaking changes are additive, not silent

  contact:
    name: neyialiyorlar project
servers:
  - url: http://localhost:5000/api/v1
    description: Local development server
paths:
...
  /query:
    post:
      summary: Multi-metric matrix pull (B2B batch)
      operationId: queryMetrics
      tags:
        - Metrics Values
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QueryRequest'
      responses:
        '200':
          description: Matrix of requested metrics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryResponse'
        '413':
          description: Payload too large (exceeds max_query_metrics)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    Confidence:
      type: string
      enum:
        - fresh
        - stale
        - approx
      description: |
        Data quality flag:
        - **fresh**: current value from primary source, < staleness threshold
        - **stale**: delayed or secondary source, past staleness threshold
        - **approx**: proxy metric or degraded computation
    MetricValue:
      type: object
      required:
        - metric
        - entity
        - ts
        - flag
        - tier
      properties:
        metric:
          type: string
          description: Metric key (e.g. nimvi, velocity_accumulation)
        entity:
          type: string
          description: Entity/security identifier
        ts:
          type: string
          format: date-time
          description: Observation timestamp (ISO 8601 UTC)
        value:
          type: number
          nullable: true
          description: Metric value; null for gap rows (never fabricated)
        flag:
          $ref: '#/components/schemas/Confidence'
        tier:
          type: string
          enum:
            - intraday
            - daily
            - weekly
            - event
          description: Metric cadence tier
...
    QueryRequest:
      type: object
      required:
        - queries
      properties:
        queries:
          type: array
          minItems: 1
          items:
            type: object
            required:
              - metric
              - entities
            properties:
              metric:
                type: string
                description: Metric key
              entities:
                type: array
                minItems: 1
                items:
                  type: string
                description: List of entity IDs
    QueryResponse:
      type: object
      required:
        - result
      properties:
        result:
          type: array
          items:
            $ref: '#/components/schemas/MetricValue'
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
        request_id:
          type: string
          description: Request ID for tracking
        timestamp:
          type: string
          format: date-time